[QLN-20] Move component out
