+++ src/services/quiz.service.ts
@@ -1,3 +1,5 @@
+"use server"
+
 import { getAPIServerURL } from "@/lib/utils"
 import QuizBank, { TAPIQuizResponse } from "@/types/QuizBank"
 
@@ -62,26 +64,3 @@ export async function copyQuizBankToClassroom(
   })
 }
 
-export async function deleteQuizBank(token: string, quizbankId: string) {
-  const url = getAPIServerURL(`/api/quizbank/${quizbankId}`)
-
-  const options = {
-    method: "DELETE",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-  }
-  return fetch(url, options).then(async (res) => {
-    if (!res?.ok) {
-      return {
-        isSuccess: false,
-        message: res.statusText,
-      }
-    }
-    console.log(res.json())
-    return {
-      isSuccess: true,
-    }
-  })
-}
